test_project/requirements.txt
django==4.1.3
django-nested-admin